dashboard specs - common mocks and includes to a shared file
